test updated
git-svn-id: https://kdlibcpp.svn.codeplex.com/svn@90892 9b283d60-5439-405e-af05-b73fd8c4d996

+++ kdlib/tests/kdlibtest/basefixture.h
@@ -15,7 +15,7 @@ public:
     kdlib::PROCESS_ID  StartTargetappWithParam( const std::wstring& cmdline ) 
     {
         std::vector<wchar_t>   buffer(0x1000);
-        DWORD   len = GetCurrentDirectory( buffer.size(), &buffer[0] );
+        DWORD   len = GetCurrentDirectory( static_cast<DWORD>(buffer.size()), &buffer[0] );
 
         std::wstring  path = std::wstring( &buffer[0], len );
         path += L"\\targetapp.exe";
